1. Hibernate orm  2. Bean validation  3. 2 level cache (infinispan)  4. Inhiritance hibernate

+++ src/main/java/com/zhmenko/author/mapper/AuthorMapper.java
@@ -22,7 +22,7 @@ public abstract class AuthorMapper {
         return authorEntity.getSecondName() + " " + authorEntity.getFirstName() + " " + authorEntity.getThirdName();
     }
 
-    public AuthorEntity authorIdToAuthorEntity(Integer authorId) {
+    public AuthorEntity authorIdToAuthorEntity(Long authorId) {
         AuthorEntity authorEntity = new AuthorEntity();
         authorEntity.setId(authorId);
         return authorEntity;
