+++ src/main/java/com/zhmenko/exception/UserNotFoundException.java
@@ -6,7 +6,7 @@ import java.text.MessageFormat;
  * Custom exception class for handling cases when a user does not exist.
  */
 public class UserNotFoundException extends EntityNotFoundException {
-    public UserNotFoundException(int id) {
+    public UserNotFoundException(Long id) {
         super(MessageFormat.format("User with id {0} not found!", id));
     }
